python/requirements.txt
Pillow>=10.0.0
plotly>=5.0.0

